pyproject.toml
[tool.poetry]
name = "warehouse_service_test_task_bristol"
version = "0.1.0"
description = "FastAPI service for warehouse state"
authors = ["Alexey Shakov <alexei_96@inbox.ru>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
sqlalchemy = ">=2.0.40,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
pre-commit = ">=4.2.0,<5.0.0"
pydantic-settings = ">=2.8.1,<3.0.0"
motor = ">=3.7.0,<4.0.0"
backoff = ">=2.2.1,<3.0.0"
aiokafka = ">=0.12.0,<0.13.0"
fastapi-cache2 = "^0.2.2"
redis = "^5.2.1"
typer = { extras = ["all"], version = "^0.15.2" }
prometheus-fastapi-instrumentator = "^7.1.0"

[tool.poetry.scripts]
warehouse-cli = "src.cli:app"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.11.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
